Delete protection

+++ app/Http/Controllers/PostController.php
@@ -31,6 +31,9 @@ class PostController extends Controller
 	public function getDeletePost($post_id)
 	{
 		$post = Post::where('id', $post_id)->first();
+		if (Auth::user() != $post->user) {
+			return redirect()->back();
+		}
 		$post->delete();
 		return redirect()->route('dashboard')->with(['message' => 'Successfully deleted']);
 	}
